backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
pandas==2.0.3
psycopg2-binary==2.9.6
pyjwt==2.7.0
requests==2.31.0
sqlalchemy==2.0.17
